+++ src/Firebase/Request.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kreait\Firebase;
+
+interface Request extends \JsonSerializable
+{
+}