+++ tests/Unit/Auth/ApiClientTest.php
@@ -8,6 +8,7 @@ use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
 use Kreait\Firebase\Auth\ApiClient;
 use Kreait\Firebase\Exception\AuthException;
+use Kreait\Firebase\Request\CreateUser;
 use Kreait\Firebase\Tests\UnitTestCase;
 use Psr\Http\Message\RequestInterface;
 
@@ -39,7 +40,7 @@ class ApiClientTest extends UnitTestCase
             ->willThrowException(new RequestException('Foo', $request->reveal()));
 
         $this->expectException(AuthException::class);
-        $this->client->signupNewUser();
+        $this->client->createUser(CreateUser::new());
     }
 
     public function testCatchThrowable()
@@ -50,6 +51,6 @@ class ApiClientTest extends UnitTestCase
             ->willThrowException(new \Exception());
 
         $this->expectException(AuthException::class);
-        $this->client->signupNewUser();
+        $this->client->createUser(CreateUser::new());
     }
 }